+++ CHANGELOG.md
@@ -17,14 +17,28 @@ Security in case of vulnerabilities.
 ## Unreleased
 
 ### Added
-- Added Project Significant History Dates `/projects/{projectId}/date-history`
-- Added `/projects/{projectId}/external-contacts` page
+- Added `Conversion` and `Transfer` task list with their statuses.
 
 ### Changed
 
 ### Fixed
 
-See the [full commit history](https://github.com/DFE-Digital/complete-conversions-transfers-changes/compare/production-2025-08-05.608...main) for everything awaiting release
+See the [full commit history](https://github.com/DFE-Digital/complete-conversions-transfers-changes/compare/production-2025-08-12.625...main) for everything awaiting release
+
+---
+
+## [1.12.0](https://github.com/DFE-Digital/complete-conversions-transfers-changes/releases/tag/production-2025-08-12.625) - 2025-08-12
+
+### Added
+- Added Project Significant History Dates `/projects/{projectId}/date-history`
+- Added `/projects/{projectId}/external-contacts` page
+- Added `/privacy` page
+
+### Changed
+
+### Fixed
+
+See the [full commit history](https://github.com/DFE-Digital/complete-conversions-transfers-changes/compare/production-2025-08-05.608...production-2025-08-12.625) for everything in the release
 
 ---
 
@@ -77,6 +91,10 @@ See the [full commit history](https://github.com/DFE-Digital/complete-conversion
 ### Added
 - New route `/projects/all/reports`
 - Add reports landing page under All projects > Reports (`/projects/all/reports`)
+- New route `/projects/{projectId}/date-history`
+- New route `/projects/{projectId}/date-history/new`
+- New route `/projects/{projectId}/date-history/reason`
+- New route `/projects/{projectId}/date-history/reasons/later`
 
 ### Changed  
 - Add a redirect from projects/{id} to projects/{id}/tasks